create test file for SoC_OCV
